Move test_properties out of test_package
