+++ storage_devices/disk.py
@@ -9,7 +9,7 @@ from test_utils.size import Size, Unit
 from test_tools import disk_utils
 from storage_devices.partition import Partition
 from storage_devices.device import Device
-from test_package.test_properties import TestProperties
+from core.test_properties import TestProperties
 import re
 import time
