+++ test_package/example/example_test.py
@@ -6,7 +6,7 @@
 import pytest
 from test_tools.disk_utils import Filesystem
 from test_utils.size import Size, Unit
-from test_package.test_properties import TestProperties
+from core.test_properties import TestProperties
 from test_package.conftest import base_prepare
 from test_utils.filesystem.file import File
 from test_utils.filesystem.directory import Directory
