+++ test_package/load/test_load.py
@@ -7,7 +7,7 @@
 import pytest
 from api.cas import casadm, casadm_parser
 from test_package.conftest import base_prepare
-from test_package.test_properties import TestProperties
+from core.test_properties import TestProperties
 from storage_devices.disk import DiskType
 from test_utils.size import Size, Unit
